pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triage-agent"
version = "0.1.0"
description = "GitHub issue triage automation pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ricardo Henriques" }]
dependencies = [
  "requests>=2.31.0",
  "scikit-learn>=1.3.0",
  "joblib>=1.3.0",
  "numpy>=1.24.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["triage*"]
